python/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
exceptiongroup==1.2.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
mysql-connector-python==8.2.0
numpy==1.26.2
pandas==2.1.4
protobuf==4.21.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
